requirements.txt
alpaca-py
pandas
TA-Lib
requests
langchain
langchain-community
langchain-core
streamlit>=1.28.0
altair>=4.0,<6.0
tzdata
